temp_clone/unzipped/EKFAlpha-main/requirements.txt
aiofiles>=24.1.0
aiohttp>=3.11.0
asyncio>=3.4.3
asyncssh>=2.14.2
discord.py>=2.5.2
dnspython>=2.7.0
email-validator>=2.2.0
flask>=3.1.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.1
gunicorn>=23.0.0
motor==3.3.1
openai>=1.12.0
paramiko>=3.5.1
pillow>=10.2.0
psutil>=5.9.0
psycopg2-binary>=2.9.10
pydantic>=2.11.0
pymongo[srv]==4.6.1
python-dotenv>=1.1.0
pytz>=2025.2
requests>=2.31.0
sqlalchemy>=2.0.27
trafilatura>=2.0.0
werkzeug>=3.0.1
wtforms>=3.1.2